Refactor: Update error handling framework to use contracts
- Introduced `HasErrorCode` protocol and `ErrorCodeCarrierError` class in a new `contracts.py` file for better error management.
- Updated imports across various modules to utilize the new error handling contracts, enhancing consistency in error reporting.
