+++ app/ldap_protocol/dns/selfhosted.py
@@ -9,7 +9,7 @@ from dataclasses import asdict
 from ipaddress import IPv4Address, IPv6Address
 
 import dns.asyncresolver
-from errors.types import ErrorCodeCarrierError
+from errors.contracts import ErrorCodeCarrierError
 
 import ldap_protocol.dns.exceptions as dns_exc
 from enums import ErrorCode
