+++ app/ldap_protocol/ldap_requests/extended.py
@@ -8,6 +8,7 @@ from abc import ABC, abstractmethod
 from typing import AsyncGenerator, ClassVar
 
 from asn1 import Decoder
+from ldap_error_codes_mapping import get_error_code_from_ldap_code
 from loguru import logger
 from pydantic import BaseModel, SecretStr, SerializeAsAny
 from sqlalchemy import update
@@ -33,6 +34,7 @@ from ldap_protocol.ldap_responses import (
     ExtendedResponse,
 )
 from ldap_protocol.objects import ProtocolRequests
+from ldap_protocol.utils.error_codes import format_ldap_error_message
 from ldap_protocol.utils.queries import get_user
 
 from .base import BaseRequest
@@ -324,6 +326,10 @@ class ExtendedRequest(BaseRequest):
                 result_code=LDAPCodes.OPERATIONS_ERROR,
                 response_name=self.request_name,
                 response_value=None,
+                errorMessage=format_ldap_error_message(
+                    get_error_code_from_ldap_code(LDAPCodes.OPERATIONS_ERROR),
+                    str(err),
+                ),
             )
         else:
             yield ExtendedResponse(
